clean up the whole project
