+++ Sources/SwiftUIYouTubePlayer/Model/YouTubePlaybackQuality.swift
@@ -1,0 +1,8 @@
+public enum YouTubePlaybackQuality: String, Equatable {
+    case small = "small"
+    case medium = "medium"
+    case large = "large"
+    case hd720 = "hd720"
+    case hd1080 = "hd1080"
+    case highResolution = "highres"
+}
