+++ Sources/SwiftUIYouTubePlayer/Model/YouTubePlayerEvents.swift
@@ -1,0 +1,6 @@
+public enum YouTubePlayerEvents: String, Equatable {
+    case youTubeIframeAPIReady = "onYouTubeIframeAPIReady"
+    case ready = "onReady"
+    case statusChange = "onStateChange"
+    case playbackQualityChange = "onPlaybackQualityChange"
+}
