+++ Sources/SwiftUIYouTubePlayer/Model/YouTubePlayerStatus.swift
@@ -1,0 +1,8 @@
+public enum YouTubePlayerStatus: String, Equatable {
+    case unstarted = "-1"
+    case ended = "0"
+    case playing = "1"
+    case paused = "2"
+    case buffering = "3"
+    case queued = "4"
+}
