+++ Sources/SwiftUIYouTubePlayer/SwiftUIYouTubePlayer.swift
@@ -1,105 +1,20 @@
 import SwiftUI
 import SwiftUIWebView
 
-public enum YouTubePlayerAction: Equatable {
-    case idle,
-         loadURL(URL),
-         loadID(String),
-         loadPlaylistID(String),
-         mute,
-         unmute,
-         play,
-         pause,
-         stop,
-         clear,
-         seek(Float, Bool),
-         duration,
-         currentTime,
-         previous,
-         next
-}
-
-public enum YouTubePlayerStatus: String, Equatable {
-    case unstarted = "-1",
-         ended = "0",
-         playing = "1",
-         paused = "2",
-         buffering = "3",
-         queued = "4"
-}
-
-public enum YouTubePlayerEvents: String, Equatable {
-    case youTubeIframeAPIReady = "onYouTubeIframeAPIReady",
-         ready = "onReady",
-         statusChange = "onStateChange",
-         playbackQualityChange = "onPlaybackQualityChange"
-}
-
-public enum YouTubePlaybackQuality: String, Equatable {
-    case small = "small",
-         medium = "medium",
-         large = "large",
-         hd720 = "hd720",
-         hd1080 = "hd1080",
-         highResolution = "highres"
-}
-
 typealias YouTubePlayerParams = [String: AnyObject]
-
-public struct YouTubePlayerState: Equatable {
-    public internal(set) var ready: Bool
-    public internal(set) var status: YouTubePlayerStatus
-    public internal(set) var quality: YouTubePlaybackQuality
-    public internal(set) var duration: Double?
-    public internal(set) var currentTime: Double?
-    public internal(set) var error: Error?
-    internal var iframeReady: Bool
-    
-    public static let empty = YouTubePlayerState(ready: false,
-                                                 status: .unstarted,
-                                                 quality: .small,
-                                                 duration: nil,
-                                                 currentTime: nil,
-                                                 error: nil,
-                                                 iframeReady: false)
-    
-    public static func == (lhs: YouTubePlayerState, rhs: YouTubePlayerState) -> Bool {
-        lhs.ready == rhs.ready
-            && lhs.status == rhs.status
-            && lhs.quality == rhs.quality
-            && lhs.duration == rhs.duration
-            && lhs.currentTime == rhs.currentTime
-            && lhs.error?.localizedDescription == rhs.error?.localizedDescription
-            && lhs.iframeReady == rhs.iframeReady
-    }
-}
-
-public struct YouTubePlayerConfig {
-    let playInline: Bool
-    let allowControls: Bool
-    let showInfo: Bool
-    
-    public init(playInline: Bool,
-                allowControls: Bool,
-                showInfo: Bool) {
-        self.playInline = playInline
-        self.allowControls = allowControls
-        self.showInfo = showInfo
-    }
-    
-    public static let `default` = YouTubePlayerConfig(playInline: true,
-                                                      allowControls: false,
-                                                      showInfo: false)
-}
 
 /** Embed and control YouTube videos */
 public struct YouTubePlayer: View {
-    private static let webViewConfig = WebViewConfig(javaScriptEnabled: true,
-                                                     allowsBackForwardNavigationGestures: false,
-                                                     allowsInlineMediaPlayback: true,
-                                                     isScrollEnabled: false,
-                                                     isOpaque: true,
-                                                     backgroundColor: .clear)
+    
+    private static let webViewConfig = WebViewConfig(
+        javaScriptEnabled: true,
+        allowsBackForwardNavigationGestures: false,
+        allowsInlineMediaPlayback: true,
+        isScrollEnabled: false,
+        isOpaque: true,
+        backgroundColor: .clear
+    )
+    
     private static let playerHTML = """
         <!DOCTYPE html>
         <html>
@@ -146,9 +61,11 @@ public struct YouTubePlayer: View {
     @State private var webViewState = WebViewState.empty
     @State private var playerVars: YouTubePlayerParams = [:]
     
-    public init(action: Binding<YouTubePlayerAction>,
-                state: Binding<YouTubePlayerState>,
-                config: YouTubePlayerConfig = .default) {
+    public init(
+        action: Binding<YouTubePlayerAction>,
+        state: Binding<YouTubePlayerState>,
+        config: YouTubePlayerConfig = .default
+    ) {
         _action = action
         _state = state
         playerVars = ["playsinline": config.playInline ? "1" : "0",
@@ -158,12 +75,14 @@ public struct YouTubePlayer: View {
     }
 
     public var body: some View {
-        WebView(config: YouTubePlayer.webViewConfig,
-                action: $webViewAction,
-                state: $webViewState,
-                schemeHandlers: ["ytplayer": handleJSEvent(_:)])
-            .onChange(of: action, perform: handleAction(_:))
-//            .onChange(of: webViewState, perform: handleWebViewStateChange(_:))
+        WebView(
+            config: YouTubePlayer.webViewConfig,
+            action: $webViewAction,
+            state: $webViewState,
+            schemeHandlers: ["ytplayer": handleJSEvent(_:)]
+        )
+        .onChange(of: action, perform: handleAction(_:))
+        //.onChange(of: webViewState, perform: handleWebViewStateChange(_:))
     }
     
     private func handleAction(_ value: YouTubePlayerAction) {
@@ -266,19 +185,21 @@ public struct YouTubePlayer: View {
 
     // MARK: Player parameters and defaults
     private var defaultPlayerParams: YouTubePlayerParams {
-        ["height": "100%" as AnyObject,
-         "width": "100%" as AnyObject,
-         "events": playerCallbacks as AnyObject,
-         "playerVars": playerVars as AnyObject
+        [
+            "height": "100%" as AnyObject,
+            "width": "100%" as AnyObject,
+            "events": playerCallbacks as AnyObject,
+            "playerVars": playerVars as AnyObject
         ]
     }
-
+    
     private var playerCallbacks: YouTubePlayerParams {
-       ["onReady": "onReady" as AnyObject,
-        "onStateChange": "onStateChange" as AnyObject,
-        "onPlaybackQualityChange": "onPlaybackQualityChange" as AnyObject,
-        "onError": "onPlayerError" as AnyObject
-       ]
+        [
+            "onReady": "onReady" as AnyObject,
+            "onStateChange": "onStateChange" as AnyObject,
+            "onPlaybackQualityChange": "onPlaybackQualityChange" as AnyObject,
+            "onError": "onPlayerError" as AnyObject
+        ]
     }
 
     private func serializedJSON(_ object: AnyObject) -> String? {
@@ -289,7 +210,6 @@ public struct YouTubePlayer: View {
         }
         return string
     }
-
 
     // MARK: JS Event Handling
     private func handleJSEvent(_ eventURL: URL) {
@@ -323,38 +243,7 @@ public struct YouTubePlayer: View {
     }
 }
 
-private extension URL {
-    func queryStringComponents() -> [String: AnyObject] {
-
-        var dict = [String: AnyObject]()
-
-        // Check for query string
-        if let query = self.query {
-
-            // Loop through pairings (separated by &)
-            for pair in query.components(separatedBy: "&") {
-
-                // Pull key, val from from pair parts (separated by =) and set dict[key] = value
-                let components = pair.components(separatedBy: "=")
-                if (components.count > 1) {
-                    dict[components[0]] = components[1] as AnyObject?
-                }
-            }
-
-        }
-
-        return dict
-    }
-}
-
-private func videoIDFromYouTubeURL(_ videoURL: URL) -> String? {
-    if videoURL.pathComponents.count > 1 && (videoURL.host?.hasSuffix("youtu.be"))! {
-        return videoURL.pathComponents[1]
-    } else if videoURL.pathComponents.contains("embed") {
-        return videoURL.pathComponents.last
-    }
-    return videoURL.queryStringComponents()["v"] as? String
-}
+// MARK: - Preview
 
 struct YouTubeTest: View {
     @State private var action = YouTubePlayerAction.idle
